Safari-611.1.10/LayoutTests/js/map-foreach-calls-back-with-right-args-expected.txt

Tests to make sure we call forEach callback with right arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS called is true
PASS actual.value is 'value'
PASS actual.key is 'key'
PASS actual.map is m
PASS actual.receiver is receiver
PASS successfullyParsed is true

TEST COMPLETE
